backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
tensorflow>=2.0.0
Pillow>=8.0.0
numpy>=1.19.0
